requirements.txt
emoji
pydantic
pymupdf
celery
numpy
langchain_experimental
langchain-text-splitters
xlrd == 2.0.1
openpyxl
openai
transformers
pyarrow
fastparquet
spacy
docling
pytest-playwright

pymupdf
neo4j
elasticsearch==8.*
langchain-openai
langchain_experimental
langchain-text-splitters
langgraph
rank_bm25
faiss-cpu
transformers

protobuf>=5.27,<6
grpcio==1.71.2
grpcio-tools==1.71.2
grpcio-status==1.71.2
grpcio-reflection==1.71.2
googleapis-common-protos==1.66.0